Fixing numbers formatting (05.1f = 5 total, including "." and decimal value = 3 zero padded leading digits)

+++ hltb_integration/data.go
@@ -174,26 +174,26 @@ type HoursToComplete100Getter interface {
 }
 
 func secondsToHours(seconds int) float64 {
-	return float64(seconds) / float64(3600)
+	return float64(seconds) / 3600.0
 }
 
 func (d *Data) GetHoursToCompleteMain() string {
 	if gd := d.PageProps.Game.Data.Game; len(gd) > 0 {
-		return fmt.Sprintf("%03.1f hours", secondsToHours(gd[0].CompMain))
+		return fmt.Sprintf("%05.1f hours", secondsToHours(gd[0].CompMain))
 	}
 	return ""
 }
 
 func (d *Data) GetHoursToCompletePlus() string {
 	if gd := d.PageProps.Game.Data.Game; len(gd) > 0 {
-		return fmt.Sprintf("%03.1f hours", secondsToHours(gd[0].CompPlus))
+		return fmt.Sprintf("%05.1f hours", secondsToHours(gd[0].CompPlus))
 	}
 	return ""
 }
 
 func (d *Data) GetHoursToComplete100() string {
 	if gd := d.PageProps.Game.Data.Game; len(gd) > 0 {
-		return fmt.Sprintf("%03.1f hours", secondsToHours(gd[0].Comp100))
+		return fmt.Sprintf("%05.1f hours", secondsToHours(gd[0].Comp100))
 	}
 	return ""
 }
